requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
supabase==2.1.1
pyjwt==2.8.0
httpx==0.25.2
python-dotenv==1.0.0
google-generativeai==0.3.1

